backend_greenworld2/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
asyncpg==0.29.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic[email]==1.10.7
passlib>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
celery==5.2.7
redis==4.5.4
python-dotenv==1.0.0
python-multipart==0.0.9
#Potom
transliterate==1.10.2
alembic==1.14.1
MarkupSafe==3.0.2
Mako==1.3.9
httptools==0.6.4
pyOpenSSL==25.0.0
pyyaml==6.0.2 
watchfiles==1.0.4 
websockets==15.0
deprecated==1.2.18 
limits==4.0.1 
packaging==24.2 
slowapi==0.1.9 
wrapt==1.17.2